+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/Option.java
@@ -287,6 +287,16 @@ public enum Option {
      */
     PLAIN_DEFINITION_KEYS(null, null),
     /**
+     * For the "format" attribute, JSON Schema defines various supported values.
+     * <br>
+     * Some of those data-types would be included if either {@link #ADDITIONAL_FIXED_TYPES} is enabled or a custom {@link SimpleTypeModule} are added.
+     * By enabling this option, only the standard built-in "format" values would be added, which is a subset of the values considered with the
+     * {@link #EXTRA_OPEN_API_FORMAT_VALUES} being enabled.
+     *
+     * @since 4.33.0
+     */
+    STANDARD_FORMATS(null, null),
+    /**
      * For the "format" attribute, JSON Schema defines various supported values. The OpenAPI specification assigns a few more of those in order to
      * differentiate between standard data types (e.g. float vs. double) and even some more fixed data types (e.g. LocalDate, LocalDateTime) if
      * {@link #ADDITIONAL_FIXED_TYPES} is enabled. By enabling this option, these extra "format" values will be set for the respective
@@ -294,7 +304,7 @@ public enum Option {
      *
      * @since 4.15.0
      */
-    EXTRA_OPEN_API_FORMAT_VALUES(null, null),
+    EXTRA_OPEN_API_FORMAT_VALUES(null, null, Option.STANDARD_FORMATS),
     /**
      * Whether as the last step of the schema generation, unnecessary "allOf" elements (i.e. where there are no conflicts/overlaps between the
      * contained sub-schemas) should be merged into one, in order to make the generated schema more readable. This also applies to manually added